+++ tests/src/test/spark330/scala/org/apache/spark/sql/rapids/suites/RapidsJsonExpressionsSuite.scala
@@ -20,23 +20,7 @@ spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.suites
 
 import org.apache.spark.sql.catalyst.expressions.JsonExpressionsSuite
-import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.rapids.utils.RapidsTestsTrait
+import org.apache.spark.sql.rapids.utils.{RapidsJsonConfTrait, RapidsTestsTrait}
 
-class RapidsJsonExpressionsSuite extends JsonExpressionsSuite with RapidsTestsTrait {
-  override def beforeAll(): Unit = {
-    super.beforeAll()
-    SQLConf.get.setConfString("spark.rapids.sql.expression.JsonTuple", "true")
-    SQLConf.get.setConfString("spark.rapids.sql.expression.GetJsonObject", "true")
-    SQLConf.get.setConfString("spark.rapids.sql.expression.JsonToStructs", "true")
-    SQLConf.get.setConfString("spark.rapids.sql.expression.StructsToJson", "true")
-  }
-
-  override def afterAll(): Unit = {
-    super.afterAll()
-    SQLConf.get.unsetConf("spark.rapids.sql.expression.JsonTuple")
-    SQLConf.get.unsetConf("spark.rapids.sql.expression.GetJsonObject")
-    SQLConf.get.unsetConf("spark.rapids.sql.expression.JsonToStructs")
-    SQLConf.get.unsetConf("spark.rapids.sql.expression.StructsToJson")
-  }
-}
+class RapidsJsonExpressionsSuite
+  extends JsonExpressionsSuite with RapidsTestsTrait with RapidsJsonConfTrait {}
